Created a loop to print 3 names and row numbers
Create three instances of the Person object using a for loop and place
each reference of the object in an array called people.
When instantiating the Person object, make sure a randomly chosen name
(using the Math.random() method) from the names array is sent to the
constructor of the person along with what row number in the HTML the
information will be displayed in.
Ex: var person = new Person(names[someName],someRow);